Refactor UI, implement welcome modal, and update README
- Refactored Sidebar component into modular sections.
- Implemented a brand color theme (Blue/Gold) using Tailwind CSS v4.
- Added a Welcome Modal for first-time visitors.
- Improved UI consistency across the application.
- Updated README.md with detailed project information.
