+++ verification/debug_console.py
@@ -1,0 +1,25 @@
+from playwright.sync_api import sync_playwright
+
+def run():
+    with sync_playwright() as p:
+        browser = p.chromium.launch()
+        page = browser.new_page()
+
+        # Capture console logs
+        page.on("console", lambda msg: print(f"Console: {msg.text}"))
+        page.on("pageerror", lambda err: print(f"Page Error: {err}"))
+
+        print("Navigating...")
+        try:
+            page.goto("http://localhost:5173", timeout=60000)
+            page.wait_for_timeout(5000)
+
+            page.screenshot(path="verification/debug_console.png")
+
+        except Exception as e:
+            print(f"Error: {e}")
+
+        browser.close()
+
+if __name__ == "__main__":
+    run()
